refactor(server): extract from SocketLifeCycle the remoteJoin logic
